feat: add validations for authorized user and form validations

+++ hw17/services/blog.service.js
@@ -16,21 +16,33 @@ exports.createBlog = async ({ title, content, author }) => {
   });
 };
 
-exports.updateBlog = async (id, title, content) => {
-  return await blogModel.findByIdAndUpdate(
-    id,
-    { title, content },
-    { new: true }
-  );
+exports.updateBlog = async (blogId, userId, title, content) => {
+  const blog = await blogModel.findById(blogId);
+  if (!blog) {
+    return null;
+  }
+  if (!blog.author.equals(userId)) {
+    const error = new Error("unauthorized");
+    error.status = 401;
+    throw error;
+  }
+  blog.title = title || blog.title;
+  blog.content = content || blog.content;
+  await blog.save();
+  return blog;
 };
 
 exports.deleteBlog = async (userId, blogId) => {
-  const deletedBlog = await blogModel.findByIdAndDelete(blogId);
-  if (deletedBlog) {
-    await userModel.findByIdAndUpdate(userId, {
-      $pull: { blogs: deletedBlog._id },
-    });
+  const blog = await blogModel.findById(blogId);
+  if (!blog) {
+    return null;
   }
-
-  return deletedBlog;
+  if (!blog.author.equals(userId)) {
+    const error = new Error("unauthorized");
+    error.status = 401;
+    throw error;
+  }
+  await blogModel.findByIdAndDelete(blogId);
+  await userModel.findByIdAndUpdate(userId, { $pull: { blogs: blogId } });
+  return blog;
 };
